pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "henk"
version = "0.3.0"
description = "Henk — Persoonlijke AI Orchestrator"
requires-python = ">=3.11"
dependencies = [
    "typer>=0.9.0",
    "anthropic>=0.40.0",
    "chromadb>=0.4.0",
    "python-frontmatter>=1.0.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0",
    "rich>=13.0.0",
    "requests>=2.31.0",
]

[project.scripts]
henk = "henk.cli:app"

[tool.pytest.ini_options]
testpaths = ["tests"]
